spruced up bootstrap java

+++ bootstrap_java.sh
@@ -1,26 +1,40 @@
 #!/bin/bash
 
-INSTALL_ROOT=$HOME/Java
+INSTALL_ROOT=/opt/java
 
-mkdir -p $INSTALL_ROOT
+if [[ ! -d $INSTALL_ROOT ]]
+then
+  sudo mkdir -p $INSTALL_ROOT
+fi
+
+sudo chown $(whoami): $INSTALL_ROOT
 
 pushd $INSTALL_ROOT
-  curl -L -b "oraclelicense=a" http://download.oracle.com/otn-pub/java/jdk/6u45-b06/jdk-6u45-linux-x64.bin -O
-  chmod u+x jdk-6u45-linux-x64.bin
-  ./jdk-6u45-linux-x64.bin
-  rm jdk-6u45-linux-x64.bin
   for version in 7u75-b13 8u40-b26; do
-      curl -L -b "oraclelicense=a" http://download.oracle.com/otn-pub/java/jdk/$version/jdk-`echo $version | cut -d \- -f 1`-linux-x64.tar.gz -O
-      tar xvf jdk-`echo $version | cut -d \- -f 1`-linux-x64.tar.gz
-      rm jdk-`echo $version | cut -d \- -f 1`-linux-x64.tar.gz
+      java_pkg=jdk-`echo $version | cut -d \- -f 1`-linux-x64.tar.gz
+      if [[ ! -e $java_pkg ]]
+      then
+        curl -L -b "oraclelicense=a" http://download.oracle.com/otn-pub/java/jdk/$version/jdk-`echo $version | cut -d \- -f 1`-linux-x64.tar.gz -O
+        rc=$?
+        if [[ $rc != 0 ]]
+        then
+          exit
+        fi
+        if [[ -e $java_pkg ]]
+        then
+          tar xvf $java_pkg
+          rm $java_pkg
+        else
+          exit
+        fi
+      fi
   done
 popd
 
 METAPATH='$PATH'
 for path in $(ls $INSTALL_ROOT | sort -r); do
-  METAPATH=$INSTALL_ROOT/$path/bin:$METAPATH
+  echo -ne 'export JAVA_HOME='$INSTALL_ROOT'/'$path'\n' >> ${HOME}'/.java_home'
 done
 
-echo Add the following line to the end of your .bashrc:
-echo PATH=$METAPATH
-
+## Java Home
+#[[ -s ${HOME}/.java_home ]] && source ${HOME}/.java_home
